+++ src/cpu/drvutils.h
@@ -16,6 +16,6 @@
 
 void int_to_ascii(int32_t n, char str[]);
 void reverse(char str[]);
-uint32_t strlen(char str[]);
+uint16_t strlen(char str[]);
 
 #endif
